# packages/contracts/openapi/depsis.yaml
openapi: 3.1.0
info:
  title: DEPSIS API
  version: 0.1.0-phase1
  summary: Debian tabanlı çok kullanıcılı NAS platformu — Faz 1 yüzeyi
  description: |
    Bu belge tek doğruluk kaynağıdır. `apps/web` istemcisi buradan üretilir
    (`openapi-typescript`, ADR-0002); elle yazılmış istemci tipi bulunamaz.

    Kapsam: Faz 1 uçları. Konsol, uygulama kataloğu ve uzak erişim de burada — sırasıyla ADR-0018, ADR-0019 ve ADR-0020. Üçü de cihazın DIŞINDAKİ bir şeyi yönetiyor (bir kabuk, podman, zerotier-one); DEPSIS hiçbirini paketlemez, kuruluysa yönetir ve kurulu değilse 503 döner. Konsol ayrıcalıklı ajanda ÇALIŞMAZ: §2.2'nin serbest komut yasağı duruyor.
    Faz 2-3'e aittir ve kasıtlı olarak yoktur — tanımlanmamış bir uç, "yakında" diye
    duran boş bir uçtan iyidir.

    Sistem aracısı IPC şeması burada DEĞİLDİR. ADR-0006 gereği onu Rust sahiplenir ve
    schemars ile üretir: bir güven sınırının sözleşmesini, o sınırı uygulayan taraf
    tanımlamalıdır.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /console:
    post:
      tags: [console]
      summary: Yönetici konsolu oturumu aç
      description: |
        ADR-0018. Üç kapı: yönetici rolü, mevcut parola, ve süre sınırı.

        Parola oturum zaten varken de isteniyor. Bir oturum, birinin açık bıraktığı bilgisayarı
        ödünç alanın sahip olduğu şeydir; bir kabuk ise cihazın tamamıdır. Parola değişikliğinde
        aynı gerekçeyle aynı şey yapılıyor.

        Konsol AYRICALIKLI AJANDA ÇALIŞMAZ. Kendi systemd birimi olan ayrı bir servis, varsayılan
        olarak ayrıcalıksız bir kullanıcıyla. §2.2'nin "aracı serbest komut kabul etmemeli"
        kuralı bozulmadı; aracı bu işin hiçbir yerinde yok.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OpenConsoleRequest' }
      responses:
        '201':
          description: Oturum açıldı
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsoleSession' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '503':
          # Konsol servisi kapatılabilir bir birim — `systemctl disable depsis-console` özelliği
          # tamamen kapatır. Kapalıyken 500 değil 503 dönmek, "bozuk" ile "kapalı" arasındaki
          # farkı koruyor.
          $ref: '#/components/responses/Problem'
    get:
      tags: [console]
      summary: Açık konsol oturumları
      description: Yalnız yöneticiler. Bir yöneticinin "şu an kimin kabuğu açık" sorusu.
      responses:
        '200':
          description: Oturumlar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConsoleSessionPage' }
        '403': { $ref: '#/components/responses/Problem' }
  /console/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [console]
      summary: Oturumu kapat
      responses:
        '204': { description: Kapatıldı }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /console/{id}/input:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [console]
      summary: Konsola tuş gönder
      description: |
        Gövde ham metindir, satır değil: bir terminal karakter karakter çalışır ve Ctrl-C,
        sekme tamamlama, ok tuşları hep bayt dizileridir. Satır bazlı bir uç, terminal olmazdı.

        Denetim için satırlar ayrıca toplanır (`console_commands`), ama bu uç noktanın işi değil —
        konsol servisi tam bir satır gördüğünde yazar.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsoleInput' }
      responses:
        '204': { description: İletildi }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '410':
          description: Oturum kapanmış — boşta kalma süresi ya da üst sınır.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
  /console/{id}/resize:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [console]
      summary: Pencere boyutunu bildir
      description: |
        Pty'nin satır/sütun sayısı. Bildirilmezse `top` ve `vim` yanlış boyutta çizer ve satır
        kaydırma bozulur — bir terminalin gerçekten terminal olup olmadığı buradan anlaşılır.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsoleResize' }
      responses:
        '204': { description: Bildirildi }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /console/{id}/stream:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [console]
      summary: Konsol çıktısı (Server-Sent Events)
      description: |
        WebSocket değil. Nest'te SSE çerçevede zaten var, WS ise bir gateway ve ek bağımlılık
        istiyor; bir terminalin gecikme bütçesi insan yazma hızı ve SSE bunun için fazlasıyla
        yeterli. Tek yönlü olması eksiklik değil — giriş zaten ayrı ve küçük bir istek.

        Her olay bir `data:` satırı; içerik base64, çünkü terminal çıktısı UTF-8 sınırlarını
        ortadan bölebilen ham bayttır ve SSE metin taşır.
      responses:
        '200':
          description: Olay akışı
          content:
            text/event-stream:
              schema: { type: string }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: depsis_session
      description: |
        Sunucu tarafı oturum. JWT kullanılmaz — güvenlik olayında tüm oturumların iptal
        edilebilmesi gerekiyor ve stateless token bunu ucuz yapmıyor (ADR-0009).
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail:
          type: string
          description: Secret, dosya içeriği veya çağıranın görmeye yetkili olmadığı yol içermez.
        instance: { type: string }
        code:
          type: string
          description: Kapalı küme. Kaynak tanım packages/contracts/src/errors.ts.
        correlationId:
          type: string
          description: Her yanıtta ve aynı isteğin her log satırında bulunur (§14).
        errors:
          type: array
          items: { $ref: '#/components/schemas/ProblemFieldError' }
        retryAfter: { type: integer, minimum: 0 }
...
    ProblemFieldError:
      type: object
      required: [pointer, code, message]
      properties:
        pointer: { type: string }
        code: { type: string }
        message: { type: string }
...
    OpenConsoleRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          maxLength: 1024
          description: Mevcut parola. Oturum varken de isteniyor (ADR-0018).
        cols: { type: integer, minimum: 20, maximum: 500, default: 80 }
        rows: { type: integer, minimum: 5, maximum: 200, default: 24 }
    ConsoleSession:
      type: object
      required: [id, username, privileged, openedAt, idleTimeoutSeconds, maxAgeSeconds]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        privileged:
          type: boolean
          description: |
            Konsolun root olarak çalışıp çalışmadığı. Varsayılan false ve açmak birim dosyasını
            elle düzenlemeyi gerektirir — bir onay kutusuyla açılabilen root kabuk, kaza ile
            açılabilen root kabuktur.
        openedAt: { type: string, format: date-time }
        idleTimeoutSeconds: { type: integer }
        maxAgeSeconds: { type: integer }
    ConsoleSessionPage:
      type: object
      required: [items, available]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ConsoleSession' }
        available:
          type: boolean
          description: Konsol servisi çalışıyor mu. Kapalıysa liste boş döner ve bu bir hata değil.
    ConsoleInput:
      type: object
      required: [data]
      properties:
        data:
          type: string
          maxLength: 8192
          description: base64. Terminal girdisi ham bayttır ve UTF-8 sınırlarını bölebilir.
    ConsoleResize:
      type: object
      required: [cols, rows]
      properties:
        cols: { type: integer, minimum: 20, maximum: 500 }
        rows: { type: integer, minimum: 5, maximum: 200 }
...